Bonsai update procedure

+++ setup_default_config.py
@@ -38,8 +38,10 @@ def delete_untracked_files(iblrig_params_path):
     packages = IBLRIG_FOLDER / 'Bonsai' / 'Packages'
     vid = packages / 'Bonsai.Video.2.4.0-preview'
     viddes = packages / 'Bonsai.Video.Design.2.4.0-preview'
-    shutil.rmtree(vid)
-    shutil.rmtree(viddes)
+    if vid.exists():
+        shutil.rmtree(vid)
+    if viddes.exists():
+        shutil.rmtree(viddes)
 
 
 ################################################################################
